+++ cpp/src/server/Server.cpp
@@ -136,9 +136,6 @@ Server::Daemonize() {
 
 int
 Server::Start() {
-//    server::Metrics::GetInstance().Init();
-//    server::Metrics::GetInstance().exposer_ptr()->RegisterCollectable(server::Metrics::GetInstance().registry_ptr());
-//    server::Metrics::GetInstance().Init();
 
     if (daemonized_) {
         Daemonize();